Make optimisations.
The structure of 'wrappers' field used to look like this:
Map {
  'eventName' => Map { [Function: origin] => [Function: wrapper] },
  'otherEvent' => Map { [Function: otherOrigin] => [Function: otherWrapper] },
  ...
}

But it made the work harder, especially in the 'listeners' method:
listeners(eventName) {
  const { events, wrappers } = this;
  const listeners = events.get(eventName);
  const callbacks = wrappers.get(eventName);
  const reversed = reverseMap(callbacks);
  if (!listeners) return [];
  const result = [];
  for (const listener of listeners) {
    const isWrapped = reversed.has(listener);
    const output = isWrapped ? reversed.get(listener) : listener;
    result.push(output);
  }
  return result;
}
We needed the 'reverseMap' function to get all the listeners. Here you can see that we are iterating over all listeners from events, which means we just need to determine if the listener is wrapped.

So I propose to make mixins to all wrapped functions with an 'origin' field that references the original listener. I really don't like mixins, but since we work with wrapped functions only inside the program abstraction and don't expose them outside, I think this is a normal approach that will simplify the code and make it more efficient.

And now 'wrappers' field looks like:
Map {
  [Function: origin] => [Function: wrapper],
  [Function: otherOrigin] => [FUnction: otherWrapper],
  ...
}